+++ Procfile
@@ -1,1 +1,1 @@
-discordbot: python discordbot.py
+web: gunicorn app:app --log-file -
