+++ requirements.txt
@@ -1,2 +1,3 @@
 discord.py==1.2.3
 Flask==1.0.3
+gunicorn==19.9.0
